using portatble method of storing a pointer in an integral type

+++ alloc-inl.h
@@ -292,7 +292,7 @@ struct TRK_obj {
 extern struct TRK_obj* TRK[ALLOC_BUCKETS];
 extern u32 TRK_cnt[ALLOC_BUCKETS];
 
-#define TRKH(_ptr) (((((u32)(_ptr)) >> 16) ^ ((u32)(_ptr))) % ALLOC_BUCKETS)
+#define TRKH(_ptr) (((((uintptr_t)(_ptr)) >> 16) ^ ((uintptr_t)(_ptr))) % ALLOC_BUCKETS)
 
 /* Adds a new entry to the list of allocated objects. */
